+++ src/main/java/de/uni_mannheim/informatik/dws/wdi/ExerciseIdentityResolution/Comparators/author/BookAuthorComparatorMongeElkan.java
@@ -48,7 +48,7 @@ public class BookAuthorComparatorMongeElkan implements Comparator<Book, Attribut
         }
 
         // postprocessing
-        int postSimilarity = 1;
+        double postSimilarity = 1;
         if (maxSimilarity <= 0.3) {
             postSimilarity = 0;
         }
